pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "browser-research-agent"
version = "0.1.0"
description = "Browser automation research agent with evidence extraction and source scoring."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "click>=8.1.7",
  "pydantic>=2.8.2",
  "playwright>=1.47.0",
  "python-dateutil>=2.9.0",
  "rich>=13.8.0"
]

[project.optional-dependencies]
api = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6"
]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.8",
  "mypy>=1.11.2",
  "types-python-dateutil>=2.9.0.20241206"
]

[project.scripts]
browser-research-agent = "browser_research_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/browser_research_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
